+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 0.0.4(2022-08-14) wait release
 
 - add option of `AhkNekoHelp.snippets.Command`
-- add register Command of `ListIncludeTree`
+- add Command of `List #Include Tree`
 - fix highlight of `#include exp.ahk`
+- fix highlight of [multi-line](https://www.autohotkey.com/docs/Scripts.htm#continuation)
 - style: config use markdown
 
 ## 0.0.3(2022-08-13)